Improve readibility of the code using date() options
PHP date() function[1] accept lots of possible options[2] to return the
representation of a date, so instead of using regular `w` option which
is a numeric representation of the day of the week and add a comment to
display which day it is, we can use `l` option to display the actual
name directly.

Same with April fool day. Instead of `1 == date('j') && 4 == date('n')`,
we can use the option `F jS` which displays the date in actual English:
`April 1st`.

I've then removed comments about day, because I believe they are no
longer needed with these changes.

References:
[1] https://www.php.net/manual/en/function.date
[2] https://www.php.net/manual/en/datetime.format.php